refactor(file-upload-item): refactor progress bar
