+++ .changeset/sharp-fireants-push.md
@@ -1,0 +1,6 @@
+---
+'@alfalab/core-components-shared': minor
+'@alfalab/core-components': minor
+---
+
+Добавлена функция `humanFileSize`, с помощью которой можно переводить байты в человеко-читаемый вид
