+++ .changeset/shiny-terms-decide.md
@@ -1,0 +1,8 @@
+---
+'@alfalab/core-components-file-upload-item': patch
+'@alfalab/core-components': patch
+---
+
+- Рефакторинг шкалы прогресса, изменена работа с mask-image на обычное заполнение svg path
+- Функция `humanFileSize` вынесена в модуль `shared`
+
