+++ .changeset/tall-buses-yawn.md
@@ -1,0 +1,6 @@
+---
+'@alfalab/core-components-file-upload-item-v1': patch
+'@alfalab/core-components': patch
+---
+
+Функция `humanFileSize` вынесена в модуль `shared`
