+++ packages/file-upload-item-v1/package.json
@@ -13,6 +13,7 @@
         "@alfalab/core-components-icon-button": "^7.0.0",
         "@alfalab/core-components-link": "^6.0.0",
         "@alfalab/core-components-spinner": "^5.0.0",
+        "@alfalab/core-components-shared": "^1.0.0",
         "@alfalab/icons-glyph": "^2.247.0",
         "classnames": "^2.5.1",
         "tslib": "^2.4.0"
