+++ packages/file-upload-item-v1/tsconfig.json
@@ -14,6 +14,8 @@
             "@alfalab/core-components-link/*": ["../link/src/*"],
             "@alfalab/core-components-screenshot-utils": ["../screenshot-utils/src"],
             "@alfalab/core-components-screenshot-utils/*": ["../screenshot-utils/src/*"],
+            "@alfalab/core-components-shared": ["../shared/src"],
+            "@alfalab/core-components-shared/*": ["../shared/src/*"],
             "@alfalab/core-components-spinner": ["../spinner/src"],
             "@alfalab/core-components-spinner/*": ["../spinner/src/*"],
             "@alfalab/core-components-test-utils": ["../test-utils/src"],
@@ -24,6 +26,7 @@
         { "path": "../icon-button/tsconfig.build.json" },
         { "path": "../link/tsconfig.build.json" },
         { "path": "../screenshot-utils/tsconfig.build.json" },
+        { "path": "../shared/tsconfig.build.json" },
         { "path": "../spinner/tsconfig.build.json" },
         { "path": "../test-utils/tsconfig.build.json" }
     ]
