+++ packages/file-upload-item/src/__image_snapshots__/file-upload-item-uploading-file-snap.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9ce70aafa8a2d6229f6e71c4c0e9923c54ffc3b970f30ad9a117009c675c7dbc
-size 4682
+oid sha256:5b86ed2bf38c16ee8888db0d4a6f96ad249a36e1c9a1c771c445f06d564e2692
+size 5647
