--- packages/file-upload-item/src/components/status-control/components/status-control-icon/index.ts
@@ -1,1 +1,0 @@
-export { StatusControlIcon } from './status-control-icon';
