--- packages/file-upload-item/src/components/status-control/components/status-control-progress-bar/index.ts
@@ -1,1 +1,0 @@
-export { StatusControlProgressBar } from './status-control-progress-bar';
