+++ packages/file-upload-item/src/components/status-control/extension-icon/index.module.css
