--- packages/file-upload-item/src/components/status-control/index.ts
@@ -1,1 +1,0 @@
-export { StatusControl } from './status-control';
