+++ packages/file-upload-item/tsconfig.json
@@ -14,6 +14,8 @@
             "@alfalab/core-components-icon-view/*": ["../icon-view/src/*"],
             "@alfalab/core-components-screenshot-utils": ["../screenshot-utils/src"],
             "@alfalab/core-components-screenshot-utils/*": ["../screenshot-utils/src/*"],
+            "@alfalab/core-components-shared": ["../shared/src"],
+            "@alfalab/core-components-shared/*": ["../shared/src/*"],
             "@alfalab/core-components-test-utils": ["../test-utils/src"],
             "@alfalab/core-components-test-utils/*": ["../test-utils/src/*"],
             "@alfalab/core-components-typography": ["../typography/src"],
@@ -24,6 +26,7 @@
         { "path": "../icon-button/tsconfig.build.json" },
         { "path": "../icon-view/tsconfig.build.json" },
         { "path": "../screenshot-utils/tsconfig.build.json" },
+        { "path": "../shared/tsconfig.build.json" },
         { "path": "../test-utils/tsconfig.build.json" },
         { "path": "../typography/tsconfig.build.json" }
     ]
